requirements.txt
pyTelegramBotAPI>=4.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
